--- a/.github/workflows/unittest.yml
+++ b/.github/workflows/unittest.yml
@@ -29,7 +29,7 @@ jobs:
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Test with pytest
       run: |
-        coverage run --source alerce/ -m pytest tests/
+        coverage run --source lc_correction/ -m pytest tests/
         coverage xml
     - name: Codecov
       uses: codecov/codecov-action@v1