Some minor optimizations for `lc_correction.compute`

1. `apply_mag_stats`
- Getting the first and last detection once, calling `loc[idxmin]` several times was the most costly part.
- Using unique then removing nan from rfids, this way the second operation is a lot quicker.
- Instead of looking for the index idxmin, se va por la posición del valor más pequeño/grande argmin/argmax.

Test times: 7.31s -> 2.99s (4.32s faster)

2. `do_dmdt_df`
- `dmdt`:  Returning a `pd.DataFrame` instead of 3 series, this way we can access the row once.
- `do_dmdt`:  
  - Changed the signature to accept only one dataframe (join of non_detection and mag_stats). 
  - Getting magstats from the first `non_detection`.
  - Avoiding using loc several times.
- `do_dmdt_df`:
  - Using join with non_detections and magstats.
  - Passing `dt_min` from one function to another.

Test times: 6.85s -> 3.29s (3.56s faster)

3. `apply_object_stats_df`
- `apply_objstats_from_correction`: Same as `apply_mag_stats`, avoiding calling `loc[idxmin]`.
- `apply_objstats_from_magstats`: Using fid as index and getting the series with loc.

Test times: 3.65 -> 2.48 (1.17s faster)

Total time: 9s

Real time reduced ~5s (From 1000alerts/30s 1000alerts/25s)

Running on our test bench, the `do_dmdt_df` and `apply_mag_stats` still costly, but faster. 
![image](https://user-images.githubusercontent.com/4682459/95878822-035c7a80-0d4c-11eb-85e3-8f03afaabae0.png)
Test bench: 34390 alerts in 843 sec = 40.79 alerts/sec.